Refactor task_schedulers as batch_schedulers
