+++ snorkel/end_model/batch_schedulers/__init__.py
@@ -1,0 +1,4 @@
+from .sequential_scheduler import SequentialScheduler
+from .shuffled_scheduler import ShuffledScheduler
+
+batch_schedulers = {"sequential": SequentialScheduler, "shuffled": ShuffledScheduler}
